Define the site menu in just one place
It is probably better to define the site menu in the configuration file,
instead of spreading it over the content files (especially if we have to
order the menu items by using the weight parameter). The only
disadvantage is that the menu isn't automatically updated when you
remove a section or fixed page from the content.