Limit the number of posts on the home page
The home page and the blog page were pretty similar: they both displayed
all posts. If we also have content on the home page, it might be better
to be able to only display the latest posts on the front page, and have
the visitors check the /blog/ page if they want to access the complete
archive.

You can add a latestpostscount param in config.toml. If this variable is
set to 0 or unset, the behavior is the same as before, so we don't break
existing sites that upgrade the theme.