Organize content in arbitrary sections
Hugo expects the content to be organized in [arbitrary sections][1]. We
can keep the structure of the demo site if we use the same default
template for list and single content, and override for the blog list and
index pages. This will get rid of the unnecessary fixed directory in the
generated output, arbitrary sections (optionally containing other
items), as well as automatically generated menus.

[1]: http://gohugo.io/overview/source-directory/