css: Handle long names for pygments
By default, rst2html uses long pygment class names in generated HTML.
While this is configurable using the '--syntax-highlight=short' flag,
hugo does not currently do this. Work around this while we wait for the
fix by embedding both long and short names in the sample CSS file.

These names are based on the comments found in file and tested locally.
They can also be seen in the source [1].

[1] https://bitbucket.org/birkenfeld/pygments-main/src/7941677/pygments/token.py

Signed-off-by: Stephen Finucane <stephen@that.guru>